requirements.txt
pytest==7.4.3
pytest-playwright==0.4.3
playwright==1.40.0
python-dotenv==1.0.0
pytest-html==4.1.1
pytest-xdist==3.3.1
# Enhanced reporting options
allure-pytest==2.13.2
pytest-json-report==1.5.0
pytest-reportlog==0.3.0
pytest-metadata==3.0.0 